Move user model to own file
