+++ packages/authentication_repository/lib/src/models/user.dart
@@ -1,0 +1,25 @@
+import 'package:equatable/equatable.dart';
+
+/// User model.
+///
+/// [User.empty] represents an unauthenticated user.
+class User extends Equatable {
+  final String id;
+  final String email;
+  final bool emailIsVerified;
+
+  const User({
+    required this.id,
+    required this.email,
+    required this.emailIsVerified,
+  });
+
+  static const empty = User(
+    email: '',
+    id: '',
+    emailIsVerified: false,
+  );
+
+  @override
+  List<Object> get props => [id, email, emailIsVerified];
+}
